Extract shared data models to reduce duplication
- Create src/images_pipeline/core module with shared data models
- Move ProcessingConfig, ImageItem, and ProcessingResult to core/models.py
- Update all processing scripts to import from core module
- Remove duplicate data structures from serial.py, multiprocess.py, multithreads.py, asyncio_example.py
- Standardize ProcessingResult with backward compatibility for different field names
- Fix linting and type checking issues

This reduces code duplication across 4 files and centralizes data model definitions.
