+++ src/images_pipeline/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core utilities and shared components for the images pipeline."""
+
+from .models import ImageItem, ProcessingConfig, ProcessingResult
+
+__all__ = ["ProcessingConfig", "ImageItem", "ProcessingResult"]
